Recording interface - Touch devices support Fixes #3
