+++ web_client/src/components/RecitalBox/useDocumentNavigation.ts
@@ -1,0 +1,157 @@
+import { useCallback, useMemo, useState } from "react";
+
+import { Document } from "@/models";
+
+export type NavigationMoves = {
+  nextParagraph: () => void;
+  prevParagraph: () => void;
+  nextSentence: () => boolean;
+  prevSentence: () => void;
+  toParagraphSentence: (paragraphIndex: number, sentenceIndex: number) => void;
+};
+
+export type NavigationArgs = {
+  paragraph: number;
+  sentence: number;
+};
+
+export enum NavigationControls {
+  NextSentence = 1,
+  PrevSentence,
+  NextParagraph,
+  PrevParagraph,
+  GoToParagraph,
+  Record,
+}
+
+const useDocumentNavigation = (document: Document) => {
+  const [activeParagraphIndex, setActiveParagraphIndex] = useState<number>(0);
+  const [activeSentenceIndex, setActiveSentenceIndex] = useState<number>(0);
+
+  const goToParagraphSentence = useCallback(
+    (paragraphIndex: number, sentenceIndex: number) => {
+      const paragraphIndexToGoTo = Math.max(
+        0,
+        Math.min(paragraphIndex, document.paragraphs.length - 1),
+      );
+      const sentenceIndexToGoTo = Math.max(
+        0,
+        Math.min(
+          sentenceIndex,
+          document.paragraphs[paragraphIndexToGoTo].sentences.length - 1,
+        ),
+      );
+      setActiveParagraphIndex(paragraphIndexToGoTo);
+      setActiveSentenceIndex(sentenceIndexToGoTo);
+    },
+    [setActiveParagraphIndex, setActiveSentenceIndex, document],
+  );
+
+  const moveToNextSentence = useCallback(() => {
+    const activeParagraph = document.paragraphs[activeParagraphIndex];
+    const isLastParagraph =
+      activeParagraphIndex === document.paragraphs.length - 1;
+    const isLastSentence =
+      activeSentenceIndex === activeParagraph.sentences.length - 1;
+    if (isLastSentence) {
+      if (isLastParagraph) {
+        return false; // No where to go further
+      }
+      setActiveParagraphIndex((prev) => prev + 1);
+      setActiveSentenceIndex(0);
+    } else {
+      setActiveSentenceIndex((prev) => prev + 1);
+    }
+
+    return true;
+  }, [
+    setActiveParagraphIndex,
+    setActiveSentenceIndex,
+    activeParagraphIndex,
+    activeSentenceIndex,
+    document,
+  ]);
+
+  const moveToPrevSentence = useCallback(() => {
+    const isFirstParagraph = activeParagraphIndex === 0;
+    const isFirstSentence = activeSentenceIndex === 0;
+    if (isFirstSentence) {
+      if (isFirstParagraph) {
+        return; // No where to go further
+      }
+      setActiveParagraphIndex((prev) => prev - 1);
+      setActiveSentenceIndex(
+        document.paragraphs[activeParagraphIndex - 1].sentences.length - 1,
+      );
+    } else {
+      setActiveSentenceIndex((prev) => prev - 1);
+    }
+  }, [
+    setActiveParagraphIndex,
+    setActiveSentenceIndex,
+    activeParagraphIndex,
+    activeSentenceIndex,
+    document,
+  ]);
+
+  const moveToNextParagraph = useCallback(() => {
+    const isLastParagraph =
+      activeParagraphIndex === document.paragraphs.length - 1;
+    if (isLastParagraph) {
+      setActiveSentenceIndex(
+        document.paragraphs[activeParagraphIndex].sentences.length - 1,
+      );
+      return; // No where to go further
+    }
+    setActiveParagraphIndex((prev) => prev + 1);
+    setActiveSentenceIndex(0);
+  }, [
+    setActiveParagraphIndex,
+    setActiveSentenceIndex,
+    activeParagraphIndex,
+    document,
+  ]);
+
+  const moveToPrevParagraph = useCallback(() => {
+    const isFirstParagraph = activeParagraphIndex === 0;
+    setActiveSentenceIndex(0);
+    if (isFirstParagraph) {
+      return; // No where to go further
+    }
+    setActiveParagraphIndex((prev) => prev - 1);
+  }, [
+    setActiveParagraphIndex,
+    setActiveSentenceIndex,
+    activeParagraphIndex,
+    document,
+  ]);
+
+  const activeParagraph = document.paragraphs[activeParagraphIndex];
+  const activeSentence = activeParagraph.sentences[activeSentenceIndex];
+  const move = useMemo<NavigationMoves>(
+    () => ({
+      nextParagraph: moveToNextParagraph,
+      prevParagraph: moveToPrevParagraph,
+      nextSentence: moveToNextSentence,
+      prevSentence: moveToPrevSentence,
+      toParagraphSentence: goToParagraphSentence,
+    }),
+    [
+      moveToNextParagraph,
+      moveToPrevParagraph,
+      moveToNextSentence,
+      moveToPrevSentence,
+      goToParagraphSentence,
+    ],
+  );
+
+  return {
+    activeParagraphIndex,
+    activeSentenceIndex,
+    activeParagraph,
+    activeSentence,
+    move,
+  } as const;
+};
+
+export default useDocumentNavigation;
